requirements.txt
streamlit
langchain
langchain-groq
groq
pandas
python-dotenv
openai
numpy
matplotlib